requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Werkzeug==3.0.1
requests==2.31.0
arxiv==2.1.0
spacy>=3.7.0
pypdf==3.17.1
gunicorn==21.2.0
